requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
altgraph==0.17.2
appdirs==1.4.4
async-timeout==4.0.2
attrs==21.4.0
bidict==0.22.0
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
crytic-compile==0.2.2
docutils==0.16
et-xmlfile==1.1.0
Flask==2.1.2
Flask-SocketIO==4.3.1
frozenlist==1.3.0
fs==2.4.14
idna==2.10
importlib-metadata==4.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
macholib==1.15.2
MarkupSafe==2.1.1
modulegraph==0.19.2
multidict==6.0.2
openpyxl==3.0.9
prettytable==3.1.1
py2app==0.26.1
Pygments==2.6.1
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pysha3==1.0.2
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2021.3
requests==2.24.0
six==1.16.0
slither-analyzer==0.8.2
typing_extensions==4.0.1
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.7.0
